Convert EA data to OCPI format
With the extras added, it's got almost all of the information
required. The `party_id` is a guess until that gets confirmed
separatelConvert EA data to OCPI format

With the extras added, it's got almost all of the information
required. The `party_id` is a guess until that gets confirmed
separately.