Add initial Enel X support
This works when the servers are behaving, but only for the time that
the first access token is available. A middleware needs to be added
for refreshing the access token every couple of minutes so it is
able to work for longer periods of time. This also needs to parse
and normalize the charging point IDs.

In Westfield, MA there are a series of stations that are marked offline
which use a different CPO ID than the other Enel X stations, but
which otherwise appear to have been on the Enel X network.