Fix missing EV Gateway stations and pre-filter
This fixes an issue with some EV Gateway networks, including the
primary one, not including stations if a network filter is included.
Unfortunately some sub-networks, like Siemens, require a network
filter in order to retrieve stations so we need to issue both requests.
Luckily Scrapy will filter out the duplicate requests for locations
so this does not cause issues.

This also makes use of offline reverse geocoding to determine the
country and state so we can filter out stations before we go to
request their data. This is important because EV Gateway will gladly
give you every station on the planet when you ask for a 10 mile radius
so we need to filter out ones we don't care about.